Improve types tests
