+++ oxidation/libparsec/crates/core_fs/Cargo.toml
@@ -7,6 +7,10 @@ autotests = false
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[[test]]
+name = "integration"
+path = "tests/mod.rs"
+
 [features]
 test-utils = []
