+++ oxidation/libparsec/crates/core_fs/tests/mod.rs
@@ -1,0 +1,3 @@
+// Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 (eventually AGPL-3.0) 2016-present Scille SAS
+
+mod test_regex;
