+++ oxidation/libparsec/crates/types/Cargo.toml
@@ -46,13 +46,12 @@ fnmatch-regex = "0.2.0"
 sha2 = "0.10.6"
 
 [dev-dependencies]
-libparsec_tests_fixtures = { path = "../tests_fixtures" }
+libparsec_tests_types = { path = "../tests_types" }
 
 pretty_assertions = "1.3.0"
 serde_test = "1.0.159"
 hex-literal = "0.3.3"
 rstest_reuse = "0.4.0"
-rstest = "0.17.0"
 
 [build-dependencies]
 glob = "0.3.0"
