+++ oxidation/libparsec/crates/types/src/regex.rs
@@ -88,100 +88,3 @@ impl Display for Regex {
         )
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::{fs, path::Path};
-
-    use crate::regex::Regex;
-    use libparsec_tests_fixtures::{tmp_path, TmpPath};
-    use rstest::rstest;
-
-    #[rstest]
-    #[case::base("*.rs\n*.py", "base.tmp")]
-    #[case::trim_whitespace("  *.rs\n   *.py   ", "trim_whitespace.tmp")]
-    #[case::empty_lines("*.rs\n\n\n*.py", "empty_lines.tmp")]
-    #[case::trim_whitespace_and_empty_lines(
-        "   *.rs\n\n  \n\n*.py  ",
-        "trim_whitespace_and_empty_lines.tmp"
-    )]
-    #[case::ignore_comment(
-        "# This contains patterns\n## yes\n   *.rs\n\n  \n\n*.py  ",
-        "ignore_comment.tmp"
-    )]
-    fn from_pattern_file_content(
-        tmp_path: TmpPath,
-        #[case] file_content: &str,
-        #[case] filename: &str,
-    ) {
-        let filepath = tmp_path.join(filename);
-        fs::write(&filepath, file_content).unwrap();
-        let regex = Regex::from_file(&filepath).expect("Regex should be valid");
-
-        assert!(regex.is_match("file.py"));
-        assert!(regex.is_match("file.rs"));
-        assert!(!regex.is_match("stories.txt"));
-    }
-
-    #[test]
-    fn load_default_pattern_file() {
-        let regex = Regex::from_file(Path::new(
-            "../../../../parsec/core/resources/default_pattern.ignore",
-        ))
-        .expect("Load default pattern file failed");
-
-        for pattern in &[
-            "file.swp",
-            "file~",
-            "$RECYCLE.BIN",
-            "desktop.ini",
-            "shortcut.lnk",
-        ] {
-            assert!(regex.is_match(pattern), "Pattern {} should match", pattern);
-        }
-
-        for pattern in &["secrets.txt", "a.docx", "picture.png"] {
-            assert!(
-                !regex.is_match(pattern),
-                "Pattern {} should not match",
-                pattern
-            );
-        }
-    }
-
-    #[rstest]
-    #[case("*.rs", "foo.rs", "bar.py", "bar.rs")]
-    #[case("*bar*", "foobar.rs", "dummy", "foobarrrrr.py")]
-    #[case("$myf$le.*", "$myf$le.txt", "bar", "$myf$le.rs")]
-    fn test_wildcard_pattern(
-        #[case] pattern_str: &str,
-        #[case] valid_case: &str,
-        #[case] bad_base: &str,
-        #[case] other_case: &str,
-    ) {
-        let regex = Regex::from_glob_pattern(pattern_str).expect("Should be valid");
-        assert!(regex.is_match(valid_case));
-        assert!(!regex.is_match(bad_base));
-        assert!(regex.is_match(other_case));
-    }
-
-    #[rstest]
-    #[case(r"fooo?$", "foo", "fo", "fooo")]
-    fn test_regex_pattern(
-        #[case] regex_str: &str,
-        #[case] valid_case: &str,
-        #[case] bad_base: &str,
-        #[case] other_case: &str,
-    ) {
-        let regex = Regex::from_regex_str(regex_str).expect("Should be valid");
-        assert!(regex.is_match(valid_case));
-        assert!(!regex.is_match(bad_base));
-        assert!(regex.is_match(other_case));
-    }
-
-    #[test]
-    fn test_bad_regex_str_creation() {
-        let r = crate::regex::Regex::from_regex_str(r"fooo][?");
-        assert!(r.is_err());
-    }
-}
