+++ oxidation/libparsec/crates/types/tests/test_local_manifest.rs
@@ -1,16 +1,14 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 (eventually AGPL-3.0) 2016-present Scille SAS
 
 use hex_literal::hex;
-use rstest::rstest;
 use std::{
     collections::{HashMap, HashSet},
     num::NonZeroU64,
 };
 
 use libparsec_crypto::prelude::*;
+use libparsec_tests_types::{alice, rstest, timestamp, Device};
 use libparsec_types::*;
-
-use libparsec_tests_fixtures::{alice, timestamp, Device};
 
 type AliceLocalFileManifest = Box<dyn FnOnce(&Device) -> (&'static [u8], LocalFileManifest)>;
 type AliceLocalFolderManifest = Box<dyn FnOnce(&Device) -> (&'static [u8], LocalFolderManifest)>;
@@ -18,7 +16,7 @@ type AliceLocalWorkspaceManifest =
     Box<dyn FnOnce(&Device) -> (&'static [u8], LocalWorkspaceManifest)>;
 type AliceLocalUserManifest = Box<dyn FnOnce(&Device) -> (&'static [u8], LocalUserManifest)>;
 
-#[rstest]
+#[rstest::rstest]
 #[case::file_manifest(Box::new(|alice: &Device| {
     let now = "2021-12-04T11:50:43.208821Z".parse().unwrap();
     (
@@ -201,7 +199,7 @@ fn serde_local_file_manifest(
     assert_eq!(manifest2, expected);
 }
 
-#[rstest]
+#[rstest::rstest]
 fn serde_local_file_manifest_invalid_blocksize() {
     // Generated from Python implementation (Parsec v2.6.0+dev)
     // Content:
@@ -245,7 +243,7 @@ fn serde_local_file_manifest_invalid_blocksize() {
     assert!(manifest.is_err());
 }
 
-#[rstest]
+#[rstest::rstest]
 #[case::folder_manifest(Box::new(|alice: &Device| {
     let now = "2021-12-04T11:50:43.208821Z".parse().unwrap();
     (
@@ -380,7 +378,7 @@ fn serde_local_folder_manifest(
     assert_eq!(manifest2, expected);
 }
 
-#[rstest]
+#[rstest::rstest]
 #[case::workspace_manifest(Box::new(|alice: &Device| {
     let now = "2021-12-04T11:50:43.208821Z".parse().unwrap();
     (
@@ -565,7 +563,7 @@ fn serde_local_workspace_manifest(
     assert_eq!(manifest2, expected);
 }
 
-#[rstest]
+#[rstest::rstest]
 #[case::need_sync(Box::new(|alice: &Device| {
     let now = "2021-12-04T11:50:43.208821Z".parse().unwrap();
     (
@@ -916,7 +914,7 @@ fn serde_local_user_manifest(
     assert_eq!(manifest2, expected);
 }
 
-#[rstest]
+#[rstest::rstest]
 fn chunk_new() {
     let chunk = Chunk::new(1, NonZeroU64::try_from(5).unwrap());
 
@@ -932,7 +930,7 @@ fn chunk_new() {
     assert_ne!(chunk, Chunk::new(1, NonZeroU64::try_from(5).unwrap()));
 }
 
-#[rstest]
+#[rstest::rstest]
 fn chunk_evolve_as_block() {
     let chunk = Chunk::new(1, NonZeroU64::try_from(5).unwrap());
     let id = chunk.id;
@@ -979,7 +977,7 @@ fn chunk_evolve_as_block() {
     assert_eq!(err, "This chunk is not aligned");
 }
 
-#[rstest]
+#[rstest::rstest]
 fn chunk_is_block() {
     let chunk = Chunk {
         id: ChunkID::default(),
@@ -1034,7 +1032,7 @@ fn chunk_is_block() {
     assert!(block.is_block());
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_file_manifest_new(timestamp: DateTime) {
     let author = DeviceID::default();
     let parent = EntryID::default();
@@ -1057,7 +1055,7 @@ fn local_file_manifest_new(timestamp: DateTime) {
     assert_eq!(lfm.blocks.len(), 0);
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_file_manifest_is_reshaped(timestamp: DateTime) {
     let author = DeviceID::default();
     let parent = EntryID::default();
@@ -1091,7 +1089,7 @@ fn local_file_manifest_is_reshaped(timestamp: DateTime) {
     assert!(!lfm.is_reshaped());
 }
 
-#[rstest]
+#[rstest::rstest]
 #[case::empty((0, vec![]))]
 #[case::blocks((2, vec![
     BlockAccess {
@@ -1140,7 +1138,7 @@ fn local_file_manifest_from_remote(timestamp: DateTime, #[case] input: (u64, Vec
     );
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_file_manifest_to_remote(timestamp: DateTime) {
     let t1 = timestamp;
     let t2 = t1.add_us(1);
@@ -1181,7 +1179,7 @@ fn local_file_manifest_to_remote(timestamp: DateTime) {
     assert_eq!(fm.blocks, vec![block_access]);
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_file_manifest_match_remote(timestamp: DateTime) {
     let fm = FileManifest {
         author: DeviceID::default(),
@@ -1221,7 +1219,7 @@ fn local_file_manifest_match_remote(timestamp: DateTime) {
     assert!(lfm.match_remote(&fm));
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_folder_manifest_new(timestamp: DateTime) {
     let author = DeviceID::default();
     let parent = EntryID::default();
@@ -1240,7 +1238,7 @@ fn local_folder_manifest_new(timestamp: DateTime) {
     assert_eq!(lfm.remote_confinement_points.len(), 0);
 }
 
-#[rstest]
+#[rstest::rstest]
 #[case::empty((
     HashMap::new(),
     HashMap::new(),
@@ -1296,7 +1294,7 @@ fn local_folder_manifest_from_remote(
     assert_eq!(lfm.remote_confinement_points.len(), filtered);
 }
 
-#[rstest]
+#[rstest::rstest]
 #[case::empty(HashMap::new(), HashMap::new(), HashMap::new(), 0, 0, false, "")]
 #[case::children_filtered(
     HashMap::from([
@@ -1396,7 +1394,7 @@ fn local_folder_manifest_from_remote_with_local_context(
     assert_eq!(lfm.remote_confinement_points.len(), filtered);
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_folder_manifest_to_remote(timestamp: DateTime) {
     let t1 = timestamp;
     let t2 = t1.add_us(1);
@@ -1421,7 +1419,7 @@ fn local_folder_manifest_to_remote(timestamp: DateTime) {
     assert_eq!(fm.children, lfm.children);
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_folder_manifest_match_remote(timestamp: DateTime) {
     let fm = FolderManifest {
         author: DeviceID::default(),
@@ -1446,7 +1444,7 @@ fn local_folder_manifest_match_remote(timestamp: DateTime) {
     assert!(lfm.match_remote(&fm));
 }
 
-#[rstest]
+#[rstest::rstest]
 #[case::empty(HashMap::new(), HashMap::new(), HashMap::new(), 0, false, "")]
 #[case::no_data(
     HashMap::new(),
@@ -1517,7 +1515,7 @@ fn local_folder_manifest_evolve_children_and_mark_updated(
 }
 
 // TODO
-#[rstest]
+#[rstest::rstest]
 fn local_folder_manifest_apply_prevent_sync_pattern(timestamp: DateTime) {
     let prevent_sync_pattern = Regex::from_regex_str("").unwrap();
 
@@ -1550,7 +1548,7 @@ fn local_folder_manifest_apply_prevent_sync_pattern(timestamp: DateTime) {
     assert_eq!(lfm.remote_confinement_points, HashSet::new());
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_workspace_manifest_new(timestamp: DateTime) {
     let author = DeviceID::default();
     let id = EntryID::default();
@@ -1571,7 +1569,7 @@ fn local_workspace_manifest_new(timestamp: DateTime) {
     assert_eq!(lwm.speculative, speculative);
 }
 
-#[rstest]
+#[rstest::rstest]
 #[case::empty((
     HashMap::new(),
     HashMap::new(),
@@ -1628,7 +1626,7 @@ fn local_workspace_manifest_from_remote(
     assert!(!lwm.speculative);
 }
 
-#[rstest]
+#[rstest::rstest]
 #[case::empty(HashMap::new(), HashMap::new(), HashMap::new(), 0, 0, false, "")]
 #[case::children_filtered(
     HashMap::from([
@@ -1729,7 +1727,7 @@ fn local_workspace_manifest_from_remote_with_local_context(
     assert!(!lwm.speculative);
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_workspace_manifest_to_remote(timestamp: DateTime) {
     let t1 = timestamp;
     let t2 = t1.add_us(1);
@@ -1754,7 +1752,7 @@ fn local_workspace_manifest_to_remote(timestamp: DateTime) {
     assert_eq!(wm.children, lwm.children);
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_workspace_manifest_match_remote(timestamp: DateTime) {
     let wm = WorkspaceManifest {
         author: DeviceID::default(),
@@ -1779,7 +1777,7 @@ fn local_workspace_manifest_match_remote(timestamp: DateTime) {
     assert!(lwm.match_remote(&wm));
 }
 
-#[rstest]
+#[rstest::rstest]
 #[case::empty(HashMap::new(), HashMap::new(), HashMap::new(), 0, false, "")]
 #[case::no_data(
     HashMap::new(),
@@ -1849,7 +1847,7 @@ fn local_workspace_manifest_evolve_children_and_mark_updated(
 }
 
 // TODO
-#[rstest]
+#[rstest::rstest]
 fn local_workspace_manifest_apply_prevent_sync_pattern(timestamp: DateTime) {
     let prevent_sync_pattern = Regex::from_regex_str("").unwrap();
 
@@ -1882,7 +1880,7 @@ fn local_workspace_manifest_apply_prevent_sync_pattern(timestamp: DateTime) {
     assert_eq!(lwm.remote_confinement_points, HashSet::new());
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_user_manifest_new(timestamp: DateTime) {
     let author = DeviceID::default();
     let id = EntryID::default();
@@ -1901,7 +1899,7 @@ fn local_user_manifest_new(timestamp: DateTime) {
     assert_eq!(lum.speculative, speculative);
 }
 
-#[rstest]
+#[rstest::rstest]
 #[case::empty((0, vec![]))]
 #[case::last_processed_message((10, vec![]))]
 #[case::workspaces((0, vec![WorkspaceEntry::generate("alice".parse().unwrap(), "2000-01-01T00:00:00Z".parse().unwrap())]))]
@@ -1926,7 +1924,7 @@ fn local_user_manifest_from_remote(timestamp: DateTime, #[case] input: (u64, Vec
     assert_eq!(lum.workspaces, um.workspaces);
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_user_manifest_to_remote(timestamp: DateTime) {
     let t1 = timestamp;
     let t2 = t1.add_us(1);
@@ -1952,7 +1950,7 @@ fn local_user_manifest_to_remote(timestamp: DateTime) {
     assert_eq!(um.workspaces, lum.workspaces);
 }
 
-#[rstest]
+#[rstest::rstest]
 fn local_user_manifest_match_remote(timestamp: DateTime) {
     let um = UserManifest {
         author: DeviceID::default(),
