+++ oxidation/libparsec/crates/types/tests/test_time.rs
@@ -1,5 +1,8 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 (eventually AGPL-3.0) 2016-present Scille SAS
 
+use hex_literal::hex;
+
+use libparsec_tests_types::rstest;
 use libparsec_types::{DateTime, Duration};
 
 type Part = (Box<dyn Fn(DateTime) -> u32>, i64);
@@ -48,4 +51,79 @@ fn test_datetime() {
         DateTime::from_f64_with_us_precision(dt.get_f64_with_us_precision()),
         dt
     );
+}
+
+#[test]
+fn test_datetime_deserialize_has_microsecond_precision() {
+    let serialized = &hex!("d70141d86ad584cd5d4f")[..];
+    let expected_timestamp_nanos = 1638618643208820000;
+
+    let dt: DateTime = rmp_serde::from_slice(serialized).unwrap();
+    assert_eq!(dt.as_ref().timestamp_nanos(), expected_timestamp_nanos);
+
+    // Round trip
+    let serialized2 = rmp_serde::to_vec_named(&dt).unwrap();
+    assert_eq!(serialized2, serialized);
+}
+
+#[test]
+fn test_datetime_attributes() {
+    let dt = DateTime::from_rfc3339("2000-1-2T12:30:59.123456Z").unwrap();
+    assert_eq!(dt.year(), 2000);
+    assert_eq!(dt.month(), 1);
+    assert_eq!(dt.day(), 2);
+    assert_eq!(dt.hour(), 12);
+    assert_eq!(dt.minute(), 30);
+    assert_eq!(dt.second(), 59);
+    assert_eq!(dt.microsecond(), 123456);
+}
+
+#[rstest::rstest]
+#[case("2000-01-01T00:00:00Z", 0)]
+#[case("2000-01-01T00:00:00+00:00", 0)]
+#[case("2000-01-01T01:00:00+01:00", 0)]
+#[case("2000-01-01T01:25:00+01:25", 0)]
+#[case("1999-12-31T23:00:00-01:00", 0)]
+// Test with milliseconds
+#[case("2000-01-01T00:00:00.123+00:00", 123000)]
+#[case("2000-01-01T01:00:00.123+01:00", 123000)]
+#[case("2000-01-01T01:25:00.123+01:25", 123000)]
+#[case("1999-12-31T23:00:00.123-01:00", 123000)]
+// Test with microseconds
+#[case("2000-01-01T00:00:00.123456Z", 123456)]
+#[case("2000-01-01T00:00:00.123456+00:00", 123456)]
+#[case("2000-01-01T01:00:00.123456+01:00", 123456)]
+#[case("2000-01-01T01:25:00.123456+01:25", 123456)]
+#[case("1999-12-31T23:00:00.123456-01:00", 123456)]
+fn test_rfc3339_conversion(#[case] raw: &str, #[case] micro: u32) {
+    let dt = DateTime::from_rfc3339(raw);
+    assert_eq!(
+        dt,
+        Ok(DateTime::from_ymd_hms_us(2000, 1, 1, 0, 0, 0, micro).unwrap())
+    );
+
+    let dt = dt.unwrap();
+
+    let expected = match micro {
+        0 => "2000-01-01T00:00:00Z".to_owned(),
+        x if x % 1000 == 0 => format!("2000-01-01T00:00:00.{}Z", micro / 1000),
+        _ => format!("2000-01-01T00:00:00.{micro}Z"),
+    };
+    assert_eq!(dt.to_rfc3339(), expected);
+
+    // Finally cheap test on idempotence
+    assert_eq!(DateTime::from_rfc3339(&dt.to_rfc3339()), Ok(dt));
+}
+
+#[rstest::rstest]
+#[case("2000-01-01")] // Missing time part
+#[case("2000-01-01T00:00:00")] // Missing tz
+#[case("2000-01-01T00:00:00+42:00")]
+#[case("2000-01-01T01:00:")]
+#[case("2000-01-01T01")]
+#[case("2000-01-01T")]
+#[case("whatever")]
+fn test_datetime_from_rfc3339_bad_parsing(#[case] raw: &str) {
+    let ret = DateTime::from_rfc3339(raw);
+    assert!(ret.is_err());
 }
